feat: use docker-buildx when DOCKER_BUILDKIT is 1
By default, Docker builds are now handled by BuildKit (buildx), unless
it is expressly turned off by setting DOCKER_BUILDKIT to an empty
string.

This makes use of BuildKit's better handling of cache_from directives,
so now these images are not pulled down in advance of a build, as
BuildKit itself is capable of checking if they are valid caches and
pulling layers down on the fly.

This change also adds two new caching modes, set by
AUTO_DEVOPS_BUILD_CACHE.

When this is set to "inline", BuildKit will cache build metadata to the
target image registries.

When this is set to another string, this is taken as an registry
reference and this is used to cache the build in "max" mode (ie where
all layers for all build targets are cached).

When this is empty or unset, advanced BuildKit caching is not used.